Only save API credentials to config if `sslmate link` is run
If the config doesn't contain API credentials and a command besides
link is run, the user is still prompted for username and password,
but the API credentials are NOT stored in the config.  This change
is being done to avoid a potential surprise that API credentials
(which are security-sensitive) have been permanently stored on the
filesystem.