Support key rollover in `sslmate download`
If a key file named ${CN}.key.new exists, first try to download
a certificate matching its public key.  If a certificate is found,
install it, and rename ${CN}.key.new to ${CN}.key in the process.

In the future, this will allow a new key to be generated by
`sslmate reissue` but not installed until its certificate has been
reissued by the certificate authority.

If no ${CN}.key.new file exists or it exists but no certificate
matches it, `sslmate download` works as it did previously.