Fix reissue and renew commands returning error status code on success.
When reissue or renew were called, the command line tool was exiting
with a status code of 1, even if the command successfully completed.
Since this is considered a failure, this made it a little tricky to
write wrapper scripts around these commands. This fix ensures these
commands exit with a successful exit code of 0 if they were actually
successful.

(This could also be fixed by having do_wait_for_cert itself return 0,
but it seemed like having the command_* methods explicitly return 0 was
more consistent with how the other command_* methods are currently
implemented.)

Conflicts:
	bin/sslmate