Always try to re-setup approval even if server thinks it's ready
But ignore errors if server thinks it's ready.

Just because server thinks token is ready doesn't mean it really is.
For instance, Route 53 could be in the process of deleting a DNS
record. The server will see the DNS record, but it will be gone in
a minute unless the client tries to set it up again.