Add ability to reissue without generating a new key
This is useful for reissuing a cert after adding or removing subjectAltNames.

'sslmate reissue' has gained a new option, --same-key, to reissue with
the same key.

A new sub-command, 'sslmate rekey', has been introduced to reissue with
a new key (like 'sslmate reissue' without --same-key).  'sslmate reissue'
without --same-key is now deprecated in favor of 'sslmate rekey'.  If
you run 'sslmate reissue' without --same-key, you will receive a warning
but it will otherwise function in a backwards-compatible manner.  In
SSLMate 2.0, --same-key will become the default, and 'sslmate rekey'
will be the only way to reissue with a new key.