Treat errors during download as non-fatal so they don't prevent further downloads
They will still cause a message to be printed and an exit status of 1.

Previously, if you were using --all, and a single .key file had a
malformed name (e.g.  no dots), only some of your certificates would
be downloaded.  This was inconsistent with the "no certificate in your
account with this name" error, which would allow further certificates
to be downloaded.

It's definitely more robust to allow further downloads, since otherwise
a renewed certificate might not get installed.